linear option for all layer blending ino fxs
