+++ stuff/profiles/layouts/fxs/STD_inoDarkenFx.xml
@@ -2,5 +2,11 @@
   <page name="Darken Ino">
     <control>opacity</control>
     <control>clipping_mask</control>
+    <separator/>
+    <control>linear</control>
+    <vbox modeSensitive="linear" mode="1">
+      <control>gamma</control>
+      <control>premultiplied</control>
+    </vbox>
   </page>
 </fxlayout>
